+++ packages/formik-inputs/src/index.ts
@@ -1,4 +1,12 @@
 export * from "./components";
 export * from "./types";
 export * from "./useField";
-export { Theme, FieldTheme, ThemeProvider } from "react-baseline-inputs";
+
+export {
+  FieldTheme,
+  Theme,
+  ThemeProvider,
+  bootstrapTheme,
+  defaultTheme,
+  useTheme
+} from "react-baseline-inputs";
